🚚 Move consent component to widgets
